feat: add user download functionality; implement user selection and download options, enhance event display with past events carousel, and update requirements

+++ app/models.py
@@ -9,7 +9,7 @@ from sqlalchemy import desc
 user_event = db.Table(
     "user_event",
     db.Column(
-        "user_id", db.Integer, db.ForeignKey("user.register_no"), primary_key=True
+        "user_id", db.BigInteger, db.ForeignKey("user.register_no"), primary_key=True
     ),
     db.Column("event_id", db.Integer, db.ForeignKey("event.id"), primary_key=True),
 )
